remove index, rename imports
remove appleApi in ios creds

index imports
