+++ packages/expo-cli/src/appleApi/index.ts
@@ -4,4 +4,4 @@ export * from './pushKey';
 export * from './provisioningProfile';
 export * from './ensureAppExists';
 export * from './setup';
-
+export * from './createManagers';
